cleaned up cases for checking dates and added a cleanedInput method
added fixes to date checks, removed or updated comments.
added a cleanedInput method to clean the the input and return a cleanedInput string taking out any words related to the date or time before we pass it to GetDescription function.
we now get a much cleaner Description for the Task.
Changed the task to ask for a tuple of strings rather than one string, so we can return both the cleanedInput and the DateTime (or null)