added a basic UI to take user input
Added basic xaml to mainPage, and a button click event to take in a user Task.
created basic error handling for a null or whitespace entry
added code to remove "at" when this string precedes a Time, but did not go ahead as it was complicating things.
Still now need to add the create task functionality
Still may need to add checks for if it is a habit/repeat task and then this would require a frequency