Fixed duplication of Folders in listview, added TaskCreator function
Fixed duplication of Folders in listview, added TaskCreator function
Copied all code for task creation from inout to a new helper class TaskCreator, which gets called.
TODO instead of calling this from main, pass the input string through to the TaskPage and call the function from there so we can then open a dialogue in task view and edit the task after we figure out the inputs. Then we can have tick boxes for what type of task etc