latest updates
fixed drop down visiblility for frequency in Add Task
but lost functionality of the description and NOtes binding to the dialog
accidently deleted the colour converter and now redone, but colours arent quite working.
if i delete a task the listview now gets rid of all from the listView (however they are still in the sql db... i suspect the binary file is deleting all tasks in the folder, as they dont load agian until we delete the bin and db files and restart.