added view task dialog page  and button action
Still needs work displaying TaskType and Frequency Drop Down
Still needs save action to save these updates to the viewmodel and update the ListView etc.
Also on TaskPage when we add or delete a Task it removes all from the ListView.
Also when we use the InputBox these details dont bind to the dialog box
Adding a task and deleting an added task works for The SQL db, but the Listview removes some or all from view,... doesnt update.
If i delete original tasks it dissappears from listview but stays in SQL db?
Checkbox does not update in SQL